relocation of calculateWinner and whoPlayNext
