+++ pages/api/mark/[id].js
@@ -1,6 +1,8 @@
 import { v4 as uuidv4 } from 'uuid';
 import Database from '../../../db/database';
-import buildRepresentation from '../../../helpers/board';
+import { buildRepresentation, 
+         whoPlayNext,
+         calculateWinner } from '../../../helpers/board';
 
 export default async (req, res) => {
   const {
@@ -62,28 +64,3 @@ export default async (req, res) => {
   });
 }
 
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      
-      return squares[a];
-    }
-  }
-  return null;
-}
-
-function whoPlayNext(step) {
-  return step % 2 ? 'X' : 'O';
-}
